feat: Re-add missing WindowDimensions resources

+++ src/utils/window.rs
@@ -2,32 +2,32 @@
 /// It's immediatly updated when the window resize event happens.
 #[derive(Default, Debug)]
 pub struct WindowDimensions {
-    width: f32,
-    height: f32,
+    width: u32,
+    height: u32,
 }
 
 impl WindowDimensions {
-    pub fn new(screen_size: (f32, f32)) -> Self {
+    pub fn new(screen_size: (u32, u32)) -> Self {
         Self {
             width: screen_size.0,
             height: screen_size.1,
         }
     }
 
-    pub fn set(&mut self, width: f32, height: f32) {
+    pub fn set(&mut self, width: u32, height: u32) {
         self.width = width;
         self.height = height;
     }
 
-    pub fn get(self) -> (f32, f32) {
+    pub fn get(self) -> (u32, u32) {
         (self.width, self.height)
     }
 
-    pub fn width(self) -> f32 {
+    pub fn width(self) -> u32 {
         self.width
     }
 
-    pub fn height(self) -> f32 {
+    pub fn height(self) -> u32 {
         self.height
     }
 }
